+++ src/main/java/com/models/av3_estrutura_de_dados/Entities/Arvores/NosAvores/NoAvoreCompraCliente.java
@@ -3,11 +3,11 @@ package com.models.av3_estrutura_de_dados.Entities.Arvores.NosAvores;
 public class NoAvoreCompraCliente {
     private String nomeProduto;
     private double precoProduto;
-    private int idVendedor;
-    private int idCliente;
+    private Long idVendedor;
+    private Long idCliente;
     private NoAvoreCompraCliente esquerda, direita;
 
-    public NoAvoreCompraCliente(String nomeProduto, double precoProduto, int idVendedor, int idCliente) {
+    public NoAvoreCompraCliente(String nomeProduto, double precoProduto, Long idVendedor, Long idCliente) {
         this.nomeProduto = nomeProduto;
         this.precoProduto = precoProduto;
         this.idVendedor = idVendedor;
@@ -23,11 +23,11 @@ public class NoAvoreCompraCliente {
         return precoProduto;
     }
 
-    public int getIdVendedor() {
+    public Long getIdVendedor() {
         return idVendedor;
     }
 
-    public int getIdCliente() {
+    public Long getIdCliente() {
         return idCliente;
     }
